index.php
This is used in my environment for gate control. Pressing Open triggers the relay to pull, opening the gate and holding it open. Pressing "Open" again, releases the relay, allowing a weight switch to trigger the gate to close.  The css shows the state the relay is in by changing the color if relay is closed.  Pressing Close jogs the relay.